+++ src/App.js
@@ -1,6 +1,6 @@
 import Cars from './Cars.js';
-import InputView from './view/InputView.js';
-import OutputView from './view/OutputView.js';
+import InputView from './views/InputView.js';
+import OutputView from './views/OutputView.js';
 import { Console } from '@woowacourse/mission-utils';
 
 class App {
